Rework for rspec 2.5
The idea here is to move everything into an actual rails 3 test app.
This simplifies the testing harness significantly.  Currently it uses
the application Gemfile with an ignore on the Gemfile.lock, although
I may want to pull this up to the gem level.  4 out of 5 specs pass
currently for me on ruby 1.9.2